feat: Display mock information for group
Part of story #26799 associate a Tuleap project and a GitLab group

How to test: After completing the wizard and linking a Tuleap project to
a GitLab group, the "GitLab Group Link" tab in Git Administration will
show a page listing the informations of a group. All informations are
faked for now, getting them from the database will be done later. All
buttons are disabled and do nothing. The javascript asset logs a message
on the console to show it is loaded, but does nothing else.

You can edit LinkedGroupPresenter to test some cases (for example
whether the group has an avatar or not).

The project background should be correctly displayed when the project is
or is not linked to a GitLab group, in the wizard page and in the
information page.

Notes: DateHelper::timeAgoInWords() is used despite being deprecated for
two reasons:
- The mockup does not show a relative date/time component.
- It is much more efficient to compute that in PHP. We are not including
moment.js just for that.

Change-Id: Icc036f40ecc7dd7b674289db591c3cb0e1ca55dc