feat: Update branch prefix for linked GitLab group
This is part of story #26799 associate a Tuleap project and a GitLab group

The create branch prefix of a linked GitLab group can now be updated
using the REST endpoint PATCH gitlab_groups/:id with a json payload
like:

{
  "create_branch_prefix": "prefix/dev/01/"
}

The update must only be done by a Git administrator.
If the prefix is not valid, an error will be thrown.

Change-Id: Icf3f791277dd82f2244cd1b3e2a94b8e8f9308a4