chore: rename with "wizard" the Group link app
Part of story #26799 associate a Tuleap project and a GitLab group

No functional change expected: with the feature flag enabled, on a
project that is not yet linked to a GitLab group, the wizard is loaded
and can be completed.

Notes:
It is renamed because we will need a separate app to deal with the
administration screen when the project is linked to a GitLab group. In
order to help developers find their way, the wizard is named "wizard".

Change-Id: I3294559f7d8e97bb05bbf77139b108d328566fcb