chore: Introduce strategy to deal with recipient removal
The current implementation of RecipientManager is a suite of different
patterns applied one after another. Let's make it clear and easier to
grasp with chain of responsability pattern.

This commit also brings the logger a little bit deeper in the stack to
get more information on how and when users are removed from
notification.

Part of: story #10469 choose if I receive a notification when I create/update something

Change-Id: If4f0b46c469151527507c3df14dc0d0f022a5191