fix: remove "updated" in Heartbeat widget
Since ac65ed474d8f4485a1a4814ad117c982691d5597, the word "updated" is
left alone at the end of each artifacts in the Heartbeat widget. This
brings confusion, users don't know if "updated" is part of the artifact
title or not.

We loose context of the entry (has the artifact been created or
updated), ideally we would move the mention Updated/Created in the entry
metadata, but it deserves a dedicated contribution. For now we just fix
the UX.

Closes request #28851: remove "updated" in Heartbeat widget

Change-Id: I0c85487bc8798a2faa7e3a591544b1062f6137ad