Provide a `meilisearch` service in Docker Compose file
This will be needed for the development of the "remote Meilisearch"
part of the story.

It exposes a Meilisearch instance under the
`https://tuleap-web.tuleap-aio-dev.docker/meilisearch-remote/` URL.

To test, start the meilisearch service and restart the reverse-proxy
service to take into account the configuration change.
The URL https://tuleap-web.tuleap-aio-dev.docker/meilisearch-remote/health
should respond successfully.

Part of story #28835: have a full-text search backend based on Meilisearch

Change-Id: Ifc50a5da57186c19e2ac43eaf0646fa8b9b3385c