feat: Display the group link info from database
Part of story #26799 associate a Tuleap project and a GitLab group

How to test: After completing the wizard and linking a Tuleap project to
a GitLab group, the "GitLab Group Link" tab in Git Administration will
show a page with the actual group info. Previously it was all fake data.

Change-Id: I142624a004c04918630c26ffbfb1ce13abe8fc23