Use local Meilisearch instance
To test you need to have a working local Meilisearch instance
(see 98ff5915811d166c0498b0f91e2d35901f2518bf), also you will probably
want to index your existing documents with `tuleap full-text-search:index-all-pending-items`.

The metadata associated to the indexed items are still stored in the DB.
It avoids more complex setup of the index on the Meilisearch side and
make it is easy for us to determine an ID for a document.

Part of story #28835: have a full-text search backend based on Meilisearch

Change-Id: I6b808db83f049c10065fc5a0bf5b930495858bbb