feat: Unlink button is active
Part of story #26799 associate a Tuleap project and a GitLab group

How to test:
- Either complete the wizard or call the POST REST route manually to
link a Project and a Group.
- When you go to the Git Administration tab, the "Unlink" button is now
active. If you click it, a modal opens. If you confirm the modal, a REST
call is made and at the end the page reloads to show the wizard again.
The button stays inactive to avoid triggering an error while the page
reloads.
- If there is an error, it is shown in the modal feedback section.

Change-Id: Id697864051585d0519a17c311fc76fc9e3fd6534