feat: Display 'Content matches: xxx'
When an item matches the keywords but nothing is highlighted it is
because FTS detected that the content of the item is matching. It might
confuse people to have one item with highlighting match and other one
without. For the latter we now display "Content matches: <keywords>" to
better reflect this situation.

Ideally we would display an extract of the content that matches the
keywords, but let's stick with this simpler proposal first.

Part of story #26771: Index artifact fields (REST + Switch to) on updated artifacts only

Change-Id: I823799f1b020432651601685be9455ddcefdc6a1