feat: Admin UI for Meilisearch remote server
Administrator can now use the web interface to set the Meilisearch
remote server settings: url and api key (index name will be done in a
dedicated changeset).

If the platform is using a local server, then the interface is not
available.

To test see instructions on how to set up a local[0]/remote[1] server.

Part of story #28835: have a full-text search backend based on Meilisearch

[0] git #tuleap/stable/311dd7653f3d3106bc6faba5a6b33781487ad031,
    git #tuleap/stable/9e37f66221771d0c74bbf995a6c84ccb51a23844,
    git #tuleap/stable/98ff5915811d166c0498b0f91e2d35901f2518bf.
[1] git #tuleap/stable/f6fcb3ddd4883ddebaa0db26e92cb2dafe350668.

Change-Id: I71c249f35990775d112a6c38cb09bdf44630c375