fix request #28861 Display of Sprint Overview force computation of Release BurnUp
By procatively assigning the active pane with the full pane
we force the computation of the whole renderer (including
Burn Down and Burn Up) while the will never be displayed.

It's, by itself, a waste of resources because we only need
the title of the Pane. But its even worst when combined
with giant Burn Up that can lead to timeout (504) wherase
the burnup is not displayed at all.

Change-Id: I578663b760b811ad6cd91daefc829bd6b87530ba