Upgrade git:pre-receive-analyze command
Upgrade the git:pre-receive-analyze command so that it accepts a repository ID and a git object reference as parameter.
With these parameters it proceed to fetch informations on the given reference.

How to test:
- Execute the command tuleap git:pre-receive-analyze
- If you gave an existing repo ID and reference, you shouldn't have any output

Part of story #28845 have an hidden command executing custom command to determine if the reference should have been rejected

Change-Id: Ib30189098ec7a3072bad8cf21594841b2ab8f16c