fix: remove outline on Chrome on FP pages
On Chrome-like browsers, on FlamingParrot pages, due to Bootstrap rules,
there is an outline when you focus an element.

To have the same behavior on FP and BP pages, (and across different
browsers) we remove this css rule.

To test: open the modal, hit <tab> and arrow keys to focus different
elements. You should not have a white outline around the focused
element.

Part of story #28835: have a full-text search backend based on Meilisearch

Change-Id: Ib5ccdc3205b3d92cd05c0b2855ab3f665c1bc752