feat: Display excerpt for search results
To give a little more context to the search results, we now display an
extract of the document that matches the query. This is only available
for fts_meilisearch, fts_db still display a generic "Content matches…".

As we are still not sure if this is the way to go, this can be
deactivated via a feature flag so that we can test on real data with or
without the feature. It is enabled by default.

To deactivate:
tuleap config-set feature_flag_fts_meilisearch_display_matching_content 0

To activate:
tuleap config-set feature_flag_fts_meilisearch_display_matching_content 1

This feature and the extra margin added to have less dense results will
be adjusted later.

Part of story #28835: have a full-text search backend based on Meilisearch

Change-Id: If52306bf7baea57f119e01634c969acca2042db0