feat: Display status badge for artifacts
Given a tracker have a semantic status defined, and this semantic is
based on static values,
When one of its artifact is displayed in the SwitchTo modal, either in
recent items or in search results,
Then a badge is displayed to its right.

And if the static values have decorators based on tlp palette colors,
Then the badge reflects that color (else it default to "secondary").

Part of story #26771: Index artifact fields (REST + Switch to) on updated artifacts only

Change-Id: Ie8f4fc6e67ca986d4461f0f2f8c32780192d67cd