refactor: DOM lib to reuse common selections
Part of story #26799 associate a Tuleap project and a GitLab group

No functional change expected in the following parts:
- Keyboard shortcuts help modal, especially with plugin-specific
shortcuts like in Taskboard context
- Program management: Program view (where you plan in Program
Increments) (no change at init)
- Program management: Iterations view (no change at init)
- Git Administration, Gitlab group link tab (no change at init)

Notes: the goal of this lib is to reuse common selection operations we
do, for example reading a data-attribute on an app mount point, or
selecting an element that must exist. We often do that on app init.

Change-Id: I437ce7af91bc4b6a883c4c86ae45a9fe9bee8104