request #28870: Unreliable FRS e2e test execution
Sometimes, for unknown reasons the CI doesn't manage to find the release
name it just created in the FRS e2e tests.

To mitigate this issue the proposal is to apply the same strategy than
for program management, by reloading the page 10 times until we find
what we are looking for.

Change-Id: I3c5ee19ca1ed99513c6c0400c597df214c103a14