fix!: move subversion configuration to svnplugin
Since d6dc6a2c3322a8055054919fbebfae8bf3a261e7 the order of nginx rules
evaluation has changed and php rules are evaluated.

When a `.php` file is added in `/svnroot` path, instead of being handled
by apache/svn, it was caught by nginx rules and transfered to Tuleap.

As svn plugin is evaluated before `03-locations.php`, transfering here
solves the problem.

How to test: make a commit in a svn repository that comes from svn
legacy that adds a `.php` file. It should work.

BREAKING CHANGE: this change is opinionated as it removes a feature from
Core to move it "blindly" to a plugin. For platforms that only have svn
core and no svn plugin it will break at update. On the other hand, svn
core is supposed to be EOL since 6 months, if it break something it's
more a good news to identify platforms that are late in the party.

Closes request #28871 Cannot add new .php file in core subversion repositories

Change-Id: Ie0f1fa721ebc91b406b9edaf7651a9c0282978e1