Upgrade git:pre-receive-analyze command
Add git object content analysis in the git:pre-receive-analyze command.
The analysis is done with a piece of Rust code that is called by using PHP's FFI.
Our code simply check is the word 'Tuleap' is found in the git object content, if that is the case a rejection message in JSON format, explaining the issue is returned.
If it everything is good, the rejection message will be 'null'.
This patch also put the command behind a feature flag.

To build prereceivers.so (required for proper testing):
- `cd plugins/git/additional-packages/pre-receive-rs/`
- `nix-shell` (you need to have nix installed)
- `cargo zigbuild --target x86_64-unknown-linux-gnu.2.17 --release`

To test:
- set feature_flag_enable_pre_receive_analyze_command to 1
- launch the command with valid parameters

Part of story #28845 have an hidden command executing custom command to determine if the reference should have been rejected

Change-Id: I2a9d625d3ac1cfb4daa3a1b2cf440e8ed59c2048