fix: Whole link/project container is clickable
When you click outside of the xref+title on an item (or the project
name), but still on the container, then the browser should target the
corresponding item (or project).

Of course no change if you click directly on the link or on one of the
quick links.

Part of story #26771: Index artifact fields (REST + Switch to) on updated artifacts only

Change-Id: Id5d25b62fd4f68e660f53bbf184322c33de7e633