feat: Disabled update configuration modal
Part of story #26799 associate a Tuleap project and a GitLab group

How to test:
- Either complete the wizard or call the POST REST route manually to
link a Project and a Group.
- When you go to the Git Administration tab, the "Update configuration"
button is now active. When you click it, it opens a modal with fields to
update the group link's configuration: "allow artifact closure" and
"branch prefix". The modal submit is disabled.
- When you check the "Prefix" checkbox, it makes the "Prefix" input
required and enabled. When you uncheck it, it makes it disabled.

Notes:
Submit of the modal has been split to reduce the size of this commit. It
will be done separately.

Change-Id: I43486bf1586fe6556b600cf3d89ab083206d9dd5