feat: replace `null` by empty string in Group REST routes
Part of story #26799 associate a Tuleap project and a GitLab group

How to test:
- `"create_branch_prefix": ""` on POST route is accepted and creates an
empty prefix (NULL stored in database).
- `"create_branch_prefix": null` on POST route is now rejected with code
400. It is a breaking change but these APIs are behind a feature flag so
it is acceptable.
- `"create_branch_prefix": ""` on PATCH route is accepted and removes
the prefix (NULL stored in database).
- `"create_branch_prefix": null` is still accepted on PATCH route. It
does not update the prefix.

Notes:
Previously, we used `null` to indicate that there was no prefix on the
POST route. However it created confusion as in the PATCH route, `null`
means "no update". Passing an empty string `""` is a consistent way to
remove the branch prefix that works for both routes.

I had to allow `null` on the POST representation and do validation
manually because Restler forbids passing an empty string when a string
is required. When it's not, it allows empty string and null. I have not
found a way to allow empty string and forbid null with annotations, so I
resorted to validating manually instead.

Change-Id: If497c8669710d1ec075cdb6d7d7bc53c62e94c34