fix: non readers should not see anything
Currently users that are not readers should not see a loading state, nor
the buttons to create a new baseline or comparison (which is quite
confusing).

Since readers cannot be set via the UI yet, with this change only
project administrators can access the service. Being able to define
readers will come later.

Note: the minimal path has been chosen to restrict the access. We will
probably need an error state for this case, but we need the Design Team
for that, so… later!

Part of story #26359: define dedicated baseline administrators & readers

Change-Id: Ia7d964ae777540ebd894353631317b62252fc5c0