Add the gitlab_groups/:id/synchronize REST route
Part of story #26799 associate a Tuleap project and a GitLab group

You can test the route with the API Explorer.

It will add missing repository integration and link the newly added
integration with the Gitlab group.
If the integration already exists, only the link is done.
If a Gitlab project is removed on the Gitlab side, the integration and
the link will still exist on the Tuleap side.

Change-Id: I30a2c5a90bf0430b55253cac421eda8584d37eca