feat: Active update configuration modal
Part of story #26799 associate a Tuleap project and a GitLab group

How to test:
- Either complete the wizard or call the POST REST route manually to
link a Project and a Group.
- When you go to the Git Administration tab, click on the "Update
configuration" button and submit the modal. A REST call is made and your
changes should be taken into account.
- If there is an error, it is shown in the modal feedback section.

Notes:
- The "Unlink" modal feedback was never cleared. If there is an error
and you submit the modal again, the feedback should now be hidden before
loading.

Change-Id: I4747f98bb5f89893f9fd9ad84ef29e2c5a4e978b