fix: Polish Synchronize REST route
Part of story #26799 associate a Tuleap project and a GitLab group

No functional change expected when calling the POST synchronize REST
route.

Notes:
- We should not compare datetime values in tests unless we control them.
Passing the datetime from an upper level makes us able to inject it in
the test and avoid flaky results.
- PFUser has been moved in SynchronizeGroupLinkCommand. We can have only
one parameter that way.
- GroupTokenNotFoundFault was no longer built in production code, so
it's been removed.
- Some minor tweaking of the docblock for the REST route. IMO it's not
necessary to explain how to delete things, people who will read this
will probably be in the API Explorer and can figure that out by
themselves.
- GitlabProjectIntegrator takes IntegrateRepositoriesInGroupLinkCommand
as parameter (1 parameter instead of 4)
- GroupLinkSynchronizerTest should assert that the last sync date has
been changed (it's an important side-effect).

Change-Id: Iafd7ace472bae51b53b214bce2262369f48ef242