Bypass changeset comment permissions when doing a "full indexing"
Checking if a changeset comment is associated to some ugroups is not
useful when indexing all the changeset comments on the instance and it
hurts performance a bit since there is no caching layers.

This contribution reduces the number of SQL queries when indexing the
whole instance for both the `fts_db` and `fts_meilisearch` plugins.

Part of story #28835: have a full-text search backend based on Meilisearch

Change-Id: Ic448d52c47b7d02d3a8121c769010e47db3fa17e