Fix integration_id value stored in plugin_gitlab_group_repository_integration
Part of story #26799 associate a Tuleap project and a GitLab group

To be consistent with app, The integration_id column from
plugin_gitlab_group_repository_integration should contains the id of the
integration from the Tuleap side and not the Id of the Gitlab project.

This contribution will delete all the link between the repository integration and a group

How to test:
 - make dev-forgeupgrade
 - Synchronize the group link via the REST route.
=> integration_id column should contain the value the integration id and
not the Gitlab project id anymore.

If you create a new Group link, the value inb the DB should still be
consistent and contains the integration id also.

Change-Id: I9ae61d873d2a76a0c9cc66444b1c23e8fdd17bbc