Mark artifacts as processed in batch when doing a full-indexation
Deleting the row each time an artifact is processed appears to be
costful. Removing the processed artifacts from the list in batch
improves the performance and reduces the number of write queries the DB
has to manage.

If something bad happen and we crash before being to be able to mark the
artifacts as processed it is not that big of a deal: artifacts will be
re-indexed a second time.

Part of story #28835: have a full-text search backend based on Meilisearch

Change-Id: I29e8f64fab7579a20407de20d04fc1bad11caa73