feat(svn): Do not recompute project groups
The default block of a project was computed twice per repository, one
when permission are read and once when the SVNAccessFile is dumped.

Worst, when all the repositories of a project where dumped, the default
block (that is common to all repositories of one project) was recomputed
(twice...) for each of them.

Let's compute it once for the project and call it a day.

With this patch, computing the SVNAccessFile for all repositories of a
project drops from 52s down to 4s.

Part of: request #28892 Updating SVNAccessFile on very big projects is very time consuming

Change-Id: Ic1acd0d3c9b585fac86b51c53ea01bdd276dd0cf