feat(svn): Do not schedule events when not needed
When there are UGROUP_MODIFY in queue for processing (for a given
project) it's useless to schedule update of SVNAccessFile because a
newer UGROUP_MODIFY will do that.

With this change, only the "last" UGROUP_MODIFY will trigger such an
update.

That's helpful when someone is removed from projects with a lof of user
groups, hence this will lead to only one SVNAccessFile regeneration
instead of one per user group.

Closes request #28892 Updating SVNAccessFile on very big projects is very time consuming

Change-Id: Ie3bf4d2a2788a6e50dd437670f58e33b3df40941