feat(svn): improve SVNAccessFile computing time
It's really not necessary to recompute if a user has access to a project
each and everytime the user is present in a user group. Let's cache
that.

This change alone drop the computing time from 1mn15 down to 6s and
saves 200'000 SQL queries (on one repository, imagine when there are 20
of them...)

This change also comes with an improvement on the developers tool to add
users. Until now it was relying on an anonymous bind to do query the
existing users but when you have more than 500 users you would hit LDAP
size limit. By using an authenticated bind, sky is the (new) limit.

Part of: request #28892 Updating SVNAccessFile on very big projects is very time consuming

Change-Id: I2d91d1b3f5f94476466e1920ebd7eba998df8257