fix request #29202 Create a GitLab branch + MR leads to an uncaught JSON parse error
When you create a branch + a merge request from an artifact in a Tuleap
project with GitLab repo integrations, you should not get a JSON.parse
error.

Notes:
This was due to the fact that the create MR route does not attach a body
to its response, so there is nothing to decode. By using "post" we don't
try to read a JSON body => no error.

Change-Id: Ice643e48db42cf2e2ca42a1e29d53bec3627f3e0