refactor: build RoleAssignment from Domain
This is part of request #29215 - Clean baseline plugin ⬡

RoleAssignment is a concept belonging to the Baseline Domain.
It was built as is by adapters, and we needed to check often that
the current role assignment exists and belong to the current
baseline project before saving or retrieving them.

This refactoring makes RoleAssignment build itself from a
ProjectIdentifier and a list of user grouop ids. During the build
phase, we make sure that the RoleAssignment is valid (existing ugroup in
given project), else we throw an error.

In this contribution, I also had to refactor the Role type, because
since adapters do not instanciate the RoleAssignment themselves anymore,
the @psalm-param Role::* $role directive was no longer convenient. To
keep enforcing the usage of an existing role, Role is now an interface
implemented by RoleBaselineAdmin and RoleBaselineReader.

No functional change expected

Change-Id: Ifcb8a9f26aa73626291a1f4482a4abdfd0c98420