refactor: ISaveProjectHistory should not be in Adapters
This is part of request #29215 - Clean baseline plugin ⬡

Currently the ISaveProjectHistory is an interface defined in the Adapter
namespace. Moreover, the class implementing it has inner logic that
should belong to a Domain object. The big problem here, is that when we
want to add entries in the ProjectHistory about role assignments, we end
up calling directly the adapter of the right from the adapter of the
left without any Domain object in between.

This refactoring aims to take out the Domain logic from ProjectHistory
class and put it in a Domain object.

No functional changes expected

Change-Id: I5498ebfe1e094167e23ba26cbb11a663acdc25f4