refactor: Domain should proceed RoleAssignment updates
part of request #29215 - Clean baseline plugin ⬡

Currently, ServiceSavePermissionController has too much
responsibilities. It extracts the user groups ids from the request's
body while it should be done by a dedicated adapter. This refactoring
introduces a RoleAssignmentFromRequestExtractor which takes care of the
extraction of the ids and returns a RoleAssignmentsToUpdate.

It also introduces a domain object called RoleAssignmentsSaver whose
role is to orchestrate the RoleAssignment Project History update when
given a ProjectIdentifier and a RoleAssignmentsToUpdate.

This is the final step of the refactoring aiming to clean the hexagon on
the RoleAssignments side. Now the ServiceSavePermissionsController only
handles the request, asks the Domain to save the permissions, and
returns a Response. The domain commands the server-side adapters to save
what's needed, and voilà.

Change-Id: If9e40d3e3fcf69be42393c8bdea7ce4e058b43fc