feat: Disabled token modal
Part of story #26799 associate a Tuleap project and a GitLab group

How to test:
- Either complete the wizard or call the POST REST route manually to
link a Project and a Group.
- When you go to the Git Administration tab, the "Update token"
button is now active. When you click it, it opens a modal with a field
to update the group link's access token. The modal submit is disabled.

Notes:
Submit of the modal has been split to reduce the size of this commit. It
will be done separately.

Change-Id: I9d097cbf0b89aebbc0f2c6ac58666d75241915fe