Replace all 401 to 403 in Gitlab REST routes
Part of story #26799 associate a Tuleap project and a GitLab group

The ressource concerned should returns 403 and not 401 when the user try
to acess to a forbidden resource. It is not only about of the lack of
authencation, even if the user is authenticated and
does not have sufficient right, he should still be rejected.

Change-Id: Id33f4abff8f9a5d2254233f6effa2559c9a4d891