feat: Initiate history page in Document
As we don't really want to touch the legacy history page in Docman, a
new view is added in Document.

To not clutter too much the review, only the Logs part is implemented.
An empty table for versions is displayed for relevant items (link, file,
embedded). Display of versions, and removal action will be contributed
later.

Since this new view is not complete, it is hidden from regular browsing.
You can access it via URL:
/plugins/document/<project-name>/history/<item-id>

In order to not reinvent the wheel, the existing display has been
refactored so that we can retrieve item logs from both the legacy docman
and document front-end. There should not be any functional changes in
legacy docman.

This serie of changes is a preliminary step to be able to distinguish
changes coming from ONLYOFFICE.

Mockup design for this interface can be found at:
https://www.figma.com/file/5C1seStjHJicg42JQUDFxg/Documents?node-id=904%3A9516

Part of story #28263: edit office document stored in Document

Change-Id: Ic912f23cb080e8445b4883cc6e3be270b2d93c6f