Allow multiple independent frontend apps in plugins
No functional changes are expected.

To keep some sanity the logic of what to keep in the source tarball used
to build the RPMs has been moved in one place.
Until now, it was done in the rpm/Makefile via some exclusions in a find
command and the specfile itself. Those two solutions were not ideal to
manage the complexity we have or to debug the process. We now uses the
.gitignore files of the repo with some extra rules to remove files that
are only needed in development or to add those we need in production but
not in dev.

Core is left untouched for now.

Part of request #29233: Allow multiple frontend apps per context

Change-Id: Ib6ff2c4046b6f3c2d57d5df590cea6f245b37a8c