feat: Update LDAP login error message to something more generic
This closes request #29232 Have a more generic error message when LDAP password is in error

Given a Tuleap user linked to any LDAP user and the Tuleap instance is using the LDAP authentication,
if the LDAP password is updated or is expired, then an error message saying to use the
LDAP credentials is displayed when the user tries to login.

This message was initially created to remember users to use their LDAP credentials instead of
the Tuleap ones when a Tuleap instance switched form local authentication to LDAP authentication.

We don't have anymore the need of this specific message.
We can display the generic login error message instead.

Change-Id: I1d63b5574e2c0e853061929ff6f8a30dac00765f