fix: Duplicated backlog item when moved to top / bottom
This closes request #29236 Duplicated backlog item when moved to top / bottom

In the planning view in agiledashboard, when moving a backlog item to the bottom or to the top (using the arrow buttons),
then the item is moved but also stays at it initial place.
User have to refresh the page to remove this wrongly duplicated item.

Change-Id: I2bed1107b9f39df1bf642f93418b03ea91192ec1