fix: do not mix version id with version number
Version number is not its identifier. The latter is unique accross the
platform, the former is not (two different documents can have version
number 1).

This distinction is necessary to be able to implement REST route that
will delete a specific version.

No functional changes.

Part of story #28263: edit office document stored in Document

Change-Id: I534740122fc4dcae18199aae829492cee75c9e80