fix: wrong display for relative date first
Given user has "Relative dates display" preference set to "Relative date
first",
Then in hidden history & versions pages the display of dates shouldn't
be broken.

Took the opportunity to move the display of dates in Document in a
common component to not have to repeat ourselves and to simplify setup
of tests. The remaining occurence of direct <tlp-relative-date> usage is
in custom properties display in quicklook. Since the type to define
property value is generic for all property types (accept string | number
| null), I prefer to not clutter the DocumentRelativeDate with this
specificity.

Part of story #28263: edit office document stored in Document

Change-Id: I7299f8943b7073cdd1c482358c93ef1aa65864d2