refactor: lib to query GitLab
Part of story #26799 associate a Tuleap project and a GitLab group

No functional change expected in the wizard to link a Tuleap project to
a GitLab group.

Notes:
The "pagination" part belongs in the lib but was not extracted from now.
It can be done later.
The lib is in Git plugin because the Project integration frontend lives
there. The retrieval of the GitLab projects should also use the lib but
was not changed; it can be done later.

Change-Id: Icbd8f6b3409aedaa0c093dd1ec8751ccf8a7f8da