feat: Have a shared comment webcomponent between timeline and file-diff views
part of story #29228 Replace the current management of comments by a
WebComponent in Overview and Changes view.

This contribution introduce a webcomponent displaying pull request
comments. This webcomponent is shared between the timeline view and the
file diff view. It will allow us to have a common base to develop
threaded comments features in something else than Angular.js, so it can
be both displayed in the future Vue.js overview page and the angular
app.

How to test:
- In the overview tab:
--> Inline comments (a.k.a comments on files) have a link to the
file-diff on which they have been posted
--> Global comments (a.k.a the ones you submit using the textarea at the
bottom of the timeline) have no files displayed
--> Events are displayed (abandon, merge, rebase etc.)

- In the file-diff view
--> The comments are displayed, there are no links to the file since
they are already displayed in their own context.
--> Comments having line returns are displayed properly (not inlined)
--> When you scroll a file having comments outside the viewport,
comment placeholders are generated on the opposite side.

Change-Id: I9b1214cc1cbc6e6daa743a151b5e5b6cfaa09a35