feat: Active token modal
Part of story #26799 associate a Tuleap project and a GitLab group

How to test:
How to test:
- Either complete the wizard or call the POST REST route manually to
link a Project and a Group.
- When you go to the Git Administration tab, click on the "Update token"
button and submit the modal. Two REST calls are made and the new token
is saved (but otherwise nothing changes on the UI).
- If the token is invalid, a specific error is shown in the modal
feedback section. If there are other errors, they are shown in the modal
feedback section.

Change-Id: Ic0f4e9c58d1091573546bb79f9416629354667cf