fix: Markdown value not interpreted in Tracker table report
This closes request #29248 Markdown value not interpreted in Tracker table report

When adding a text field as a column in a Tracker table report, if an artifact
has this text field value format is set to Markdown, then the value is not interpreted in the report.
If the value is set to HTML, then the value is interpreted.

We have to be consistent and both values must be interpreted.

How to test
-----------

* Have an artifact with a Markdown value in a text field
* Go to a table report of the tracker of the artifact
* Add this field as a column of this table

=> Markdown content must be interpreted

Change-Id: Ie9ef7d251f2510a6d416d3d77aa7a123b6c18dd4
Note: there is some technical changes in docblock to fix psalm errors.