chore: synchronizer should return a Domain object
Part of story #26799 associate a Tuleap project and a GitLab group

No functional change when synchronizing a project linked to a GitLab
group.

Notes:
Since we are going to be reusing the GroupLinkSynchronizer, it should
not be tied to REST context. The Resource maps the domain object to a
representation.

Change-Id: Iab95cf39010bb660f3ad9977c04fa30772426982