Maintain the "save token" alive when user is editing the document
The token is refreshed every 5 minutes (minus a few seconds) to extends
its expiration date. This make sure that when ONLYOFFICE want to save
the document, the token it has is still valid.

The token has a lifetime of 15 minutes and we try to refresh it every
5 minutes. This should make the process resilient enough even in case of
some connectivity or server issues.

The easiest way to test this change is to set a smaller refresh interval
in `plugins/onlyoffice/scripts/onlyoffice-editor.ts` and to check the
expiration date of the token in the database is extended.

Part of story #28263: edit office document stored in Document

Change-Id: I77cab4bfed5e33886c4a7f3313e3e47e1be8da3f